+++ discord/events/integration_create.py
@@ -1,0 +1,7 @@
+from ..objects import Integration
+from ..snowflake import Snowflake
+
+
+class IntegrationCreateEvent(Integration):
+  guild_id: Snowflake
+  """ID of the guild."""